@@ -16,11 +16,11 @@ Gem::Specification.new do |gem|
   gem.test_files    = Dir['test/**/*']
   gem.require_paths = ['lib']
 
-  gem.add_runtime_dependency     'actionpack', '~> 5.0'
+  gem.add_runtime_dependency     'actionpack', '~> 6.0'
   gem.add_runtime_dependency     'jwt',        '~> 1.5'
 
-  gem.add_development_dependency 'bundler',                  '~> 1.7'
-  gem.add_development_dependency 'rails',                    '~> 5.0'
+  gem.add_development_dependency 'bundler',                  '~> 2.1'
+  gem.add_development_dependency 'rails',                    '~> 6.0'
   gem.add_development_dependency 'rake',                     '~> 10.0'
   gem.add_development_dependency 'active_model_serializers', '~> 0.10.0'
 end
