@@ -1,3 +1,3 @@
 module RapidApi
-  VERSION = "0.3.1"
+  VERSION = "0.4.0"
 end
